requirements.txt
flet>=0.9.0
bcrypt>=4.0.1
paho-mqtt>=1.6.1
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
